fold GoodCauseName component into GoodCauseDetails
