+++ smile-street-react/src/components/GoodCauseDetails/GoodCauseDetails.js
@@ -7,7 +7,8 @@ import {
   TextField,
   Button,
 } from "@material-ui/core";
-import GoodCauseName from './GoodCauseName';
+import ApolloClient, { gql } from "apollo-boost"
+import { ApolloProvider, useQuery } from "@apollo/react-hooks"
 import PageHeading from "../PageHeading/PageHeading";
 
 const useStyles = makeStyles((theme) => ({
@@ -46,15 +47,62 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
+const client = new ApolloClient({
+  uri: "https://charitybase.uk/api/graphql",
+  headers: {
+      Authorization: "Apikey 93ad424e-fcfd-4d5e-8820-32ecf2291c54",
+  },
+})
+
+const GetGoodCauseName = (props) => {
+  const classes = useStyles();
+  let charityNumber = props.number
+  const COUNT_QUERY = gql`
+    query CountCharitiesCHC($id: [ID]) {
+      CHC {
+        getCharities(filters: {id : $id}) {
+          list {
+            names { value }
+          }
+        }
+      }
+    }
+    `
+  const { loading, error, data } = useQuery(COUNT_QUERY, {
+    variables: { "id":  charityNumber }
+  });
+  let result = '';
+  if (loading) { result = "Loading..." }
+  if (error) { result = "Error :(" }
+  if (data) { 
+    result = data.CHC.getCharities.list.length ? data.CHC.getCharities.list[0].names[0].value : "No matches";
+    props.setGoodCauseName(result === "No matches" ? "" : result)
+  }
+  return (
+      <TextField
+          className={classes.root}
+          variant="outlined"
+          margin="normal"
+          id="Name of good cause"
+          label="Name of good cause"
+          value = {result}
+          fullWidth
+          readOnly
+          />
+  )
+}
+
 
 export default function GoodCauseDetails() {
   const [goodCauseNumber, setGoodCauseNumber] = useState('')
   const [goodCauseDescription, setGoodCauseDescription] = useState('')
+  const [GoodCauseName, setGoodCauseName] = useState('');
+
   const classes = useStyles();
-  
+
   const handleClick = () => console.log(
     `Charity number: ${goodCauseNumber}
-    Good Cause name: my idea is to move the function from GoodCauseName into here and get rid of the component... -A
+    Good Cause name: ${GoodCauseName}
     Description: ${goodCauseDescription}
     `)
 
@@ -67,7 +115,7 @@ export default function GoodCauseDetails() {
             <TextField
               variant="outlined"
               margin="normal"
-              id="title"
+              id="Charity Registration Number"
               label="Charity Registration Number"
               className={classes.root}
               fullWidth
@@ -76,11 +124,13 @@ export default function GoodCauseDetails() {
               value={goodCauseNumber}
               onChange={(event) => setGoodCauseNumber(event.target.value)}
             />
-            <GoodCauseName className={classes.root} number={goodCauseNumber}/>
+            <ApolloProvider client={client}>
+              <GetGoodCauseName number={goodCauseNumber} setGoodCauseName={setGoodCauseName} />
+            </ApolloProvider>
             <TextField
               variant="outlined"
               margin="normal"
-              id="last name"
+              id="Good Cause Description"
               label="Add a few words describing your good cause"
               className={classes.root}
               value={goodCauseDescription}
@@ -89,10 +139,9 @@ export default function GoodCauseDetails() {
               multiline
               rows="6"
             />
-
             <Grid item xs={12} sm={12}>
               <Button variant="contained" className={classes.buttonColor} onClick={handleClick}>
-                Add Opportunity
+                Add Opportunities
               </Button>
             </Grid>
           </Grid>
